Added logic to modulate sending of notifications
The price of Ethereum is sampled every 30 minutes, and a notification is sent if the current price is 20% higher or lower than the pivot price. The pivot price is set to the current price if the 20% threshold is breached. We only want to be notified when a significant change occurs.